Revert "Add new paragraph to opt-in section on thank you page"
