+++ support-frontend/assets/pages/digital-subscription-checkout/digitalSubscriptionCheckout.jsx
@@ -14,7 +14,8 @@ import CustomerService from 'components/customerService/customerService';
 import SubscriptionTermsPrivacy
   from 'components/legal/subscriptionTermsPrivacy/subscriptionTermsPrivacy';
 import SubscriptionFaq from 'components/subscriptionFaq/subscriptionFaq';
-import ThankYouContent from './components/thankYou/thankYouContent';
+import ThankYouContent from './thankYouContent';
+import ThankYouPendingContent from './thankYouPendingContent';
 import CheckoutForm
   from 'pages/digital-subscription-checkout/components/digitalCheckoutForm';
 import 'stylesheets/skeleton/skeleton.scss';
@@ -65,8 +66,8 @@ const content = (
     >
       <CheckoutStage
         checkoutForm={<CheckoutForm />}
-        thankYouContentPending={<ThankYouContent isPending countryGroupId={countryGroupId} />}
-        thankYouContent={<ThankYouContent isPending={false} countryGroupId={countryGroupId} />}
+        thankYouContentPending={<ThankYouPendingContent countryGroupId={countryGroupId} />}
+        thankYouContent={<ThankYouContent countryGroupId={countryGroupId} />}
         subscriptionProduct="DigitalPack"
       />
       <ConsentBanner />
