+++ support-frontend/assets/pages/digital-subscription-checkout/digitalSubscriptionCheckout.scss
@@ -68,7 +68,7 @@
     border-bottom: 0;
     @include mq($from: desktop) {
       position: relative;
-      padding-bottom: $gu-v-spacing * 2;
+      padding-bottom: $gu-v-spacing * 4;
     }
     .svg-newsletters {
       display: block;
