+++ support-frontend/assets/pages/paper-subscription-checkout/components/thankYou.jsx
@@ -83,7 +83,6 @@ function ThankYouContent({
         <HeroPicture />
         <HeadingBlock
           overheading="Thank you for supporting our journalism!"
-          overheadingClass="--thankyou"
         >
           {isPending ?
           `Your subscription to the ${productOption} package is being processed` :
@@ -128,11 +127,6 @@ function ThankYouContent({
             />)
           }
         </Asyncronously>
-        <Text>
-          This is the option to choose if you want to hear about how to make the most of your print
-          subscription, receive a dedicated weekly email from our membership editor and get more
-          information on ways to support The Guardian.
-        </Text>
       </Content>
     </div>
   );
